mv: just move scene functions to parse
